separate files for AppSchema and logic
